Rename resnet teacher module
