+++ scripts/fine_tuning.py
@@ -31,7 +31,7 @@ from data.cifar100 import get_cifar100_loaders
 from data.imagenet32 import get_imagenet32_loaders
 
 # teacher factories
-from models.teachers.resnet_teacher import create_resnet152
+from models.teachers.resnet152_teacher import create_resnet152
 
 # partial freeze
 from modules.partial_freeze import (
